Added JSON parsing and chart generation
Merged parse_json to the main branch; generalized and integrated the functionality with the main app.py program so user-input values are used and raw data is returned.
Program now pulls coordinates from the JSON data and uses them to generate a pygal graph, which is then displayed in the browser.